code/APPLICATION/pyproject.toml
[tool.poetry]
name = "assignmnet4"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"
streamlit = "^1.40.1"
snowflake-connector-python = "^3.12.3"
pinecone = "^5.4.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
bcrypt = "^4.2.0"
google-search-results = "^2.4.2"
boto3 = "^1.35.60"
python-jose = "^3.3.0"
sentence-transformers = "^3.3.0"
openai = "0.28"
arxiv = "^2.1.3"
nltk = "^3.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
